fix: deduplicate macros/typedefs, restore UNUSED in signatures
- Remove duplicate glue/UNUSED/use from debug.h (single site: misc.h)
- Remove duplicate container_of from list.h (single site: misc.h)
- Remove duplicate must_check from sync.h (single site: misc.h)
- Remove duplicate underscore-suffixed ABI typedefs from misc.h
  (addr_t, pid_t_, uid_t_, mode_t_, off_t_, time_t_, clock_t_)
  (single site: include/ixland/linux_types.h)
- Move UNUSED() from function bodies back into parameter signatures
  in tmp.c, tty-real.c, time.c, uname.c (per UNUSED(x) contract)
- Add (int) cast for PTR_ERR(tty) return in tty.c

+++ Sources/IXLandLinuxRuntime/fs/tty.c
@@ -191,7 +191,7 @@ static int tty_device_open(int major, int minor, struct fd *fd)
         assert(driver != NULL);
         tty = tty_get(driver, major, minor);
         if (IS_ERR(tty))
-            return PTR_ERR(tty);
+            return (int)PTR_ERR(tty);
     }
 
     if (tty->driver->ops->open) {
